libaurum: Refactoring using TDD to observe the SRP
The existing AccessibleWatcher was handling event processing and app managing.
Since this violates the single responsibility principle, AccessibleWatcher is responsible for event processing only,
and the structure has been changed so that AccessibleAppManager manages apps.

To this end, we created a test with a structure that adds a MockAccessibleAppManager,
and made sure that the test is satisfied through additional implementation.

Change-Id: I8a619384ac5a7d62b89cd16c960875e56a4a1276