libaurum: Apply factory pattern for OCP, LSP using TDD
Adds generateKey method and Tests for this.

apply factory pattern for Key generates it able to be added or changed
without changing exising code.

It also no problem in the operation of the program even if an object of a child
class is replaced with an object of the parent class.

For advantaging of being easy to maintain and highly extensibel.

Change-Id: I5ee504a82f2f7cc5c3ffea9be8914034975bf015