libaurum: Apply clean code to object releated classes
The IObject interface seems to have been added in the past to define a common interface between the ScreenAnalyzer object and the Ui Object.
It seems better to delete it because there is no abstracted object other than UiObject.

Clean Code was applied to the fact that UiObject calls getAccessibleNode even though it has Node.

Change-Id: Icf1262cd5ef69a596fa8793721219404117eef34